+++ src/main/java/com/example/SchoolLunchReport/statistics/controller/dto/response/RankMenuResponse.java
@@ -1,4 +1,4 @@
-package com.example.SchoolLunchReport.statistics.rank.controller.dto;
+package com.example.SchoolLunchReport.statistics.controller.dto.response;
 
 import com.example.SchoolLunchReport.product.food.domain.entity.Food;
 import lombok.Builder;
